refactor: Move _register_dataclass to a separate module for reusability
